server/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
werkzeug==3.0.1
sqlalchemy==2.0.23
alembic==1.13.0
PyJWT==2.8.0
cryptography==41.0.7
psycopg2-binary==2.9.9
bcrypt==4.1.1
marshmallow==3.20.1